src/requirements.txt
also==0.0.2
certifi==2019.3.9
cffi==1.11.0
chardet==3.0.4
greenlet==0.4.12
idna==2.8
PyMySQL==0.9.3
python-dotenv==0.10.1
readline==6.2.4.1
redis==3.2.1
requests==2.21.0
schedule==0.6.0
SQLAlchemy==1.3.1
urllib3==1.24.1
faker==1.0.5
